+++ modules/User/Database/Migrations/2024_01_13_141648_create_user_views_table.php
@@ -11,8 +11,8 @@ return new class extends Migration {
     public function up(): void {
         Schema::create('user_views', function (Blueprint $table) {
             $table->id();
-            $table->integer("user_id");
-            $table->integer("item_id");
+            $table->unsignedBigInteger("user_id");
+            $table->unsignedBigInteger("item_id");
             $table->tinyInteger("item_type_id");
             $table->timestamps();
         });
